Visual updates
Created "camera" to move elements based on player position. Created "hitbox" elements for player and obstacles,  lowered the hitboxes in respect to the position rectangles of the player and obstacle, creating an overlap. Changed the order that the images are drawn so that higher objects are rendered first, so lower models will appear "on top" of them.