Visual updates continued
Removed the original map with a larger PNG from the program "tiled". Created a status for the Player class that registers what graphics should be displayed. The "idle" status is when the player is not moving. The player's status also takes into account direction. The direction is set during the move() function.  Created an attack status for the player, along with a cooldown for the attacks based on the in game time when the attack was made. Fixed an issue with "_idle" being added indefinitely, and fixed an issue where "_attack" was added indefinitely.  Mapped the status to the pngs of the player. Created a way to loop through the pngs to create "animation" with the animate() method.